+++ examples/misc/engine_flags_by_name.py
@@ -2,5 +2,5 @@
 
 from senzing import SzEngineFlags
 
-RESULT = SzEngineFlags.flags_by_name()
-print(f"\n{RESULT}\n")
+result = SzEngineFlags.flags_by_name()
+print(f"\n{result}\n")