+++ examples/szconfig/delete_data_source.py
@@ -10,9 +10,10 @@ DATA_SOURCE_CODE = "TEST"
 try:
     grpc_channel = grpc.insecure_channel("localhost:8261")
     sz_abstract_factory = SzAbstractFactoryGrpc(grpc_channel)
-    sz_config = sz_abstract_factory.create_config()
-    config_handle = sz_config.create_config()
-    sz_config.delete_data_source(config_handle, DATA_SOURCE_CODE)
-    sz_config.close_config(config_handle)
+    sz_configmanager = sz_abstract_factory.create_configmanager()
+    DATA_SOURCE_CODE = "TEST"
+    sz_config = sz_configmanager.create_config_from_template()
+    RESULT = sz_config.delete_data_source(DATA_SOURCE_CODE)
+    print(f"\n{RESULT}\n")
 except SzError as err:
     print(f"\nERROR: {err}\n")