+++ examples/szconfigmanager/create_config_from_template.py
@@ -8,6 +8,7 @@ from senzing_grpc import SzAbstractFactoryGrpc
 try:
     grpc_channel = grpc.insecure_channel("localhost:8261")
     sz_abstract_factory = SzAbstractFactoryGrpc(grpc_channel)
-    sz_config = sz_abstract_factory.create_config()
+    sz_configmanager = sz_abstract_factory.create_configmanager()
+    sz_config = sz_configmanager.create_config_from_template()
 except SzError as err:
     print(f"\nERROR: {err}\n")