+++ examples/szconfigmanager/set_default_config.py
@@ -1,19 +1,28 @@
 #! /usr/bin/env python3
 
+import time
+
 import grpc
 from senzing import SzError
 
 from senzing_grpc import SzAbstractFactoryGrpc
 
-DATA_SOURCE_CODE = "NAME_OF_DATASOURCE"
-
 try:
     grpc_channel = grpc.insecure_channel("localhost:8261")
     sz_abstract_factory = SzAbstractFactoryGrpc(grpc_channel)
     sz_configmanager = sz_abstract_factory.create_configmanager()
-    DATA_SOURCE_CODE = "NAME_OF_DATASOURCE"
+
+    # Create a new config.
+
     sz_config = sz_configmanager.create_config_from_template()
-    RESULT = sz_config.add_data_source(DATA_SOURCE_CODE)
-    print(f"\n{RESULT}\n")
+    data_source_code = f"REPLACE_DEFAULT_CONFIG_ID_{time.time()}"
+    sz_config.add_data_source(data_source_code)
+
+    # Persist the new default config.
+
+    CONFIG_DEFINITION = sz_config.export()
+    CONFIG_COMMENT = "Just an example"
+    CONFIG_ID = sz_configmanager.set_default_config(CONFIG_DEFINITION, CONFIG_COMMENT)
+
 except SzError as err:
     print(f"\nERROR: {err}\n")