+++ src/senzing_grpc/pb2_grpc/szconfig_pb2.py
@@ -24,7 +24,7 @@ _sym_db = _symbol_database.Default()
 
 
 
-DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n\x0eszconfig.proto\x12\x08szconfig\"G\n\x14\x41\x64\x64\x44\x61taSourceRequest\x12\x15\n\rconfig_handle\x18\x01 \x01(\x03\x12\x18\n\x10\x64\x61ta_source_code\x18\x02 \x01(\t\"\'\n\x15\x41\x64\x64\x44\x61taSourceResponse\x12\x0e\n\x06result\x18\x01 \x01(\t\"+\n\x12\x43loseConfigRequest\x12\x15\n\rconfig_handle\x18\x01 \x01(\x03\"\x15\n\x13\x43loseConfigResponse\"\x15\n\x13\x43reateConfigRequest\"&\n\x14\x43reateConfigResponse\x12\x0e\n\x06result\x18\x01 \x01(\x03\"J\n\x17\x44\x65leteDataSourceRequest\x12\x15\n\rconfig_handle\x18\x01 \x01(\x03\x12\x18\n\x10\x64\x61ta_source_code\x18\x02 \x01(\t\"\x1a\n\x18\x44\x65leteDataSourceResponse\",\n\x13\x45xportConfigRequest\x12\x15\n\rconfig_handle\x18\x01 \x01(\x03\"&\n\x14\x45xportConfigResponse\x12\x0e\n\x06result\x18\x01 \x01(\t\".\n\x15GetDataSourcesRequest\x12\x15\n\rconfig_handle\x18\x01 \x01(\x03\"(\n\x16GetDataSourcesResponse\x12\x0e\n\x06result\x18\x01 \x01(\t\"0\n\x13ImportConfigRequest\x12\x19\n\x11\x63onfig_definition\x18\x01 \x01(\t\"&\n\x14ImportConfigResponse\x12\x0e\n\x06result\x18\x01 \x01(\x03\x32\xd3\x04\n\x08SzConfig\x12R\n\rAddDataSource\x12\x1e.szconfig.AddDataSourceRequest\x1a\x1f.szconfig.AddDataSourceResponse\"\x00\x12L\n\x0b\x43loseConfig\x12\x1c.szconfig.CloseConfigRequest\x1a\x1d.szconfig.CloseConfigResponse\"\x00\x12O\n\x0c\x43reateConfig\x12\x1d.szconfig.CreateConfigRequest\x1a\x1e.szconfig.CreateConfigResponse\"\x00\x12[\n\x10\x44\x65leteDataSource\x12!.szconfig.DeleteDataSourceRequest\x1a\".szconfig.DeleteDataSourceResponse\"\x00\x12O\n\x0c\x45xportConfig\x12\x1d.szconfig.ExportConfigRequest\x1a\x1e.szconfig.ExportConfigResponse\"\x00\x12U\n\x0eGetDataSources\x12\x1f.szconfig.GetDataSourcesRequest\x1a .szconfig.GetDataSourcesResponse\"\x00\x12O\n\x0cImportConfig\x12\x1d.szconfig.ImportConfigRequest\x1a\x1e.szconfig.ImportConfigResponse\"\x00\x42X\n\x14\x63om.senzing.sdk.grpcB\rSzConfigProtoZ1github.com/senzing-garage/sz-sdk-go-grpc/szconfigb\x06proto3')
+DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n\x0eszconfig.proto\x12\x08szconfig\"K\n\x14\x41\x64\x64\x44\x61taSourceRequest\x12\x19\n\x11\x63onfig_definition\x18\x01 \x01(\t\x12\x18\n\x10\x64\x61ta_source_code\x18\x02 \x01(\t\"B\n\x15\x41\x64\x64\x44\x61taSourceResponse\x12\x0e\n\x06result\x18\x01 \x01(\t\x12\x19\n\x11\x63onfig_definition\x18\x02 \x01(\t\"N\n\x17\x44\x65leteDataSourceRequest\x12\x19\n\x11\x63onfig_definition\x18\x01 \x01(\t\x12\x18\n\x10\x64\x61ta_source_code\x18\x02 \x01(\t\"E\n\x18\x44\x65leteDataSourceResponse\x12\x0e\n\x06result\x18\x01 \x01(\t\x12\x19\n\x11\x63onfig_definition\x18\x02 \x01(\t\"2\n\x15GetDataSourcesRequest\x12\x19\n\x11\x63onfig_definition\x18\x01 \x01(\t\"(\n\x16GetDataSourcesResponse\x12\x0e\n\x06result\x18\x01 \x01(\t\"0\n\x13VerifyConfigRequest\x12\x19\n\x11\x63onfig_definition\x18\x01 \x01(\t\"&\n\x14VerifyConfigResponse\x12\x0e\n\x06result\x18\x01 \x01(\x08\x32\xe3\x02\n\x08SzConfig\x12R\n\rAddDataSource\x12\x1e.szconfig.AddDataSourceRequest\x1a\x1f.szconfig.AddDataSourceResponse\"\x00\x12[\n\x10\x44\x65leteDataSource\x12!.szconfig.DeleteDataSourceRequest\x1a\".szconfig.DeleteDataSourceResponse\"\x00\x12U\n\x0eGetDataSources\x12\x1f.szconfig.GetDataSourcesRequest\x1a .szconfig.GetDataSourcesResponse\"\x00\x12O\n\x0cVerifyConfig\x12\x1d.szconfig.VerifyConfigRequest\x1a\x1e.szconfig.VerifyConfigResponse\"\x00\x42X\n\x14\x63om.senzing.sdk.grpcB\rSzConfigProtoZ1github.com/senzing-garage/sz-sdk-go-grpc/szconfigb\x06proto3')
 
 _globals = globals()
 _builder.BuildMessageAndEnumDescriptors(DESCRIPTOR, _globals)
@@ -33,33 +33,21 @@ if not _descriptor._USE_C_DESCRIPTORS:
   _globals['DESCRIPTOR']._loaded_options = None
   _globals['DESCRIPTOR']._serialized_options = b'\n\024com.senzing.sdk.grpcB\rSzConfigProtoZ1github.com/senzing-garage/sz-sdk-go-grpc/szconfig'
   _globals['_ADDDATASOURCEREQUEST']._serialized_start=28
-  _globals['_ADDDATASOURCEREQUEST']._serialized_end=99
-  _globals['_ADDDATASOURCERESPONSE']._serialized_start=101
-  _globals['_ADDDATASOURCERESPONSE']._serialized_end=140
-  _globals['_CLOSECONFIGREQUEST']._serialized_start=142
-  _globals['_CLOSECONFIGREQUEST']._serialized_end=185
-  _globals['_CLOSECONFIGRESPONSE']._serialized_start=187
-  _globals['_CLOSECONFIGRESPONSE']._serialized_end=208
-  _globals['_CREATECONFIGREQUEST']._serialized_start=210
-  _globals['_CREATECONFIGREQUEST']._serialized_end=231
-  _globals['_CREATECONFIGRESPONSE']._serialized_start=233
-  _globals['_CREATECONFIGRESPONSE']._serialized_end=271
-  _globals['_DELETEDATASOURCEREQUEST']._serialized_start=273
-  _globals['_DELETEDATASOURCEREQUEST']._serialized_end=347
-  _globals['_DELETEDATASOURCERESPONSE']._serialized_start=349
-  _globals['_DELETEDATASOURCERESPONSE']._serialized_end=375
-  _globals['_EXPORTCONFIGREQUEST']._serialized_start=377
-  _globals['_EXPORTCONFIGREQUEST']._serialized_end=421
-  _globals['_EXPORTCONFIGRESPONSE']._serialized_start=423
-  _globals['_EXPORTCONFIGRESPONSE']._serialized_end=461
-  _globals['_GETDATASOURCESREQUEST']._serialized_start=463
-  _globals['_GETDATASOURCESREQUEST']._serialized_end=509
-  _globals['_GETDATASOURCESRESPONSE']._serialized_start=511
-  _globals['_GETDATASOURCESRESPONSE']._serialized_end=551
-  _globals['_IMPORTCONFIGREQUEST']._serialized_start=553
-  _globals['_IMPORTCONFIGREQUEST']._serialized_end=601
-  _globals['_IMPORTCONFIGRESPONSE']._serialized_start=603
-  _globals['_IMPORTCONFIGRESPONSE']._serialized_end=641
-  _globals['_SZCONFIG']._serialized_start=644
-  _globals['_SZCONFIG']._serialized_end=1239
+  _globals['_ADDDATASOURCEREQUEST']._serialized_end=103
+  _globals['_ADDDATASOURCERESPONSE']._serialized_start=105
+  _globals['_ADDDATASOURCERESPONSE']._serialized_end=171
+  _globals['_DELETEDATASOURCEREQUEST']._serialized_start=173
+  _globals['_DELETEDATASOURCEREQUEST']._serialized_end=251
+  _globals['_DELETEDATASOURCERESPONSE']._serialized_start=253
+  _globals['_DELETEDATASOURCERESPONSE']._serialized_end=322
+  _globals['_GETDATASOURCESREQUEST']._serialized_start=324
+  _globals['_GETDATASOURCESREQUEST']._serialized_end=374
+  _globals['_GETDATASOURCESRESPONSE']._serialized_start=376
+  _globals['_GETDATASOURCESRESPONSE']._serialized_end=416
+  _globals['_VERIFYCONFIGREQUEST']._serialized_start=418
+  _globals['_VERIFYCONFIGREQUEST']._serialized_end=466
+  _globals['_VERIFYCONFIGRESPONSE']._serialized_start=468
+  _globals['_VERIFYCONFIGRESPONSE']._serialized_end=506
+  _globals['_SZCONFIG']._serialized_start=509
+  _globals['_SZCONFIG']._serialized_end=864
 # @@protoc_insertion_point(module_scope)