+++ src/senzing_grpc/szconfigmanager.py
@@ -14,9 +14,10 @@ from types import TracebackType
 from typing import Any, Dict, Type, Union
 
 import grpc
-from senzing import SzConfigManager
+from senzing import SzConfig, SzConfigManager
 
 from .pb2_grpc import szconfigmanager_pb2, szconfigmanager_pb2_grpc
+from .szconfig import SzConfigGrpc
 from .szhelpers import as_str, new_exception
 
 # Metadata
@@ -75,29 +76,34 @@ class SzConfigManagerGrpc(SzConfigManager):
     # SzConfigManager methods
     # -------------------------------------------------------------------------
 
-    def add_config(
-        self,
-        config_definition: str,
-        config_comment: str,
-    ) -> int:
-        try:
-            request = szconfigmanager_pb2.AddConfigRequest(  # type: ignore[unused-ignore]
-                config_definition=as_str(config_definition),
-                config_comment=as_str(config_comment),
-            )
-            response = self.stub.AddConfig(request)
-            return int(response.result)
-        except Exception as err:
-            raise new_exception(err) from err
-
-    def _destroy(self) -> None:
-        """Null function in the sz-sdk-python-grpc implementation."""
-
-    def get_config(self, config_id: int) -> str:
+    def create_config_from_config_id(self, config_id: int) -> SzConfig:
         try:
             request = szconfigmanager_pb2.GetConfigRequest(config_id=config_id)  # type: ignore[unused-ignore]
             response = self.stub.GetConfig(request)
-            return str(response.result)
+            config_definition = str(response.result)
+            result = SzConfigGrpc(self.channel)
+            result.import_config_definition(config_definition)
+            return result
+        except Exception as err:
+            raise new_exception(err) from err
+
+    def create_config_from_string(self, config_definition: str) -> SzConfig:
+        try:
+            result = SzConfigGrpc(self.channel)
+            result.import_config_definition(config_definition)
+            result.verify_config_definition(config_definition)
+            return result
+        except Exception as err:
+            raise new_exception(err) from err
+
+    def create_config_from_template(self) -> SzConfig:
+        try:
+            request = szconfigmanager_pb2.GetTemplateConfigRequest()  # type: ignore[unused-ignore]
+            response = self.stub.GetTemplateConfig(request)
+            config_definition = str(response.result)
+            result = SzConfigGrpc(self.channel)
+            result.import_config_definition(config_definition)
+            return result
         except Exception as err:
             raise new_exception(err) from err
 
@@ -117,16 +123,20 @@ class SzConfigManagerGrpc(SzConfigManager):
         except Exception as err:
             raise new_exception(err) from err
 
-    def _initialize(
+    def register_config(
         self,
-        instance_name: str,
-        settings: Union[str, Dict[Any, Any]],
-        verbose_logging: int = 0,
-    ) -> None:
-        """Null function in the sz-sdk-python-grpc implementation."""
-        _ = instance_name
-        _ = settings
-        _ = verbose_logging
+        config_definition: str,
+        config_comment: str,
+    ) -> int:
+        try:
+            request = szconfigmanager_pb2.RegisterConfigRequest(  # type: ignore[unused-ignore]
+                config_definition=as_str(config_definition),
+                config_comment=as_str(config_comment),
+            )
+            response = self.stub.RegisterConfig(request)
+            return int(response.result)
+        except Exception as err:
+            raise new_exception(err) from err
 
     def replace_default_config_id(self, current_default_config_id: int, new_default_config_id: int) -> None:
         try:
@@ -138,6 +148,17 @@ class SzConfigManagerGrpc(SzConfigManager):
         except Exception as err:
             raise new_exception(err) from err
 
+    def set_default_config(self, config_definition: str, config_comment: str) -> int:
+        try:
+            request = szconfigmanager_pb2.SetDefaultConfigRequest(  # type: ignore[unused-ignore]
+                config_definition=as_str(config_definition),
+                config_comment=as_str(config_comment),
+            )
+            response = self.stub.SetDefaultConfig(request)
+            return int(response.result)
+        except Exception as err:
+            raise new_exception(err) from err
+
     def set_default_config_id(self, config_id: int) -> None:
         try:
             request = szconfigmanager_pb2.SetDefaultConfigIdRequest(  # type: ignore[unused-ignore]
@@ -146,3 +167,21 @@ class SzConfigManagerGrpc(SzConfigManager):
             self.stub.SetDefaultConfigId(request)
         except Exception as err:
             raise new_exception(err) from err
+
+    # -------------------------------------------------------------------------
+    # Non-public SzConfigManagerCore methods
+    # -------------------------------------------------------------------------
+
+    def _destroy(self) -> None:
+        """Null function in the sz-sdk-python-grpc implementation."""
+
+    def initialize(
+        self,
+        instance_name: str,
+        settings: Union[str, Dict[Any, Any]],
+        verbose_logging: int = 0,
+    ) -> None:
+        """Null function in the sz-sdk-python-grpc implementation."""
+        _ = instance_name
+        _ = settings
+        _ = verbose_logging