+++ tests/szabstractfactory_test.py
@@ -3,7 +3,6 @@ from datetime import datetime
 import pytest
 from senzing import (
     SzAbstractFactory,
-    SzConfig,
     SzConfigManager,
     SzDiagnostic,
     SzEngine,
@@ -19,70 +18,63 @@ FACTORY_PARAMETERS: SzAbstractFactoryParametersGrpc = {
 }
 
 # -----------------------------------------------------------------------------
-# Testcases
+# Test cases
 # -----------------------------------------------------------------------------
 
 
-def test_create_config(sz_abstract_factory: SzAbstractFactory) -> None:
-    """Create SzConfig."""
-    actual = sz_abstract_factory.create_config()
-    assert isinstance(actual, SzConfig)
-
-
-def test_create_configmanager(sz_abstract_factory: SzAbstractFactory) -> None:
-    """Create SzConfigManager."""
-    actual = sz_abstract_factory.create_configmanager()
+def test_create_configmanager(sz_abstractfactory: SzAbstractFactory) -> None:
+    """Test SzAbstractFactory.create_configmanager()."""
+    actual = sz_abstractfactory.create_configmanager()
     assert isinstance(actual, SzConfigManager)
 
 
-def test_create_diagnostic(sz_abstract_factory: SzAbstractFactory) -> None:
-    """Create SzDiagnostic."""
-    actual = sz_abstract_factory.create_diagnostic()
+def test_create_diagnostic(sz_abstractfactory: SzAbstractFactory) -> None:
+    """Test SzAbstractFactory.create_diagnostic()."""
+    actual = sz_abstractfactory.create_diagnostic()
     assert isinstance(actual, SzDiagnostic)
 
 
-def test_create_engine(sz_abstract_factory: SzAbstractFactory) -> None:
-    """Create SzEngine."""
-    actual = sz_abstract_factory.create_engine()
+def test_create_engine(sz_abstractfactory: SzAbstractFactory) -> None:
+    """Test SzAbstractFactory.create_engine()."""
+    actual = sz_abstractfactory.create_engine()
     assert isinstance(actual, SzEngine)
 
 
-def test_create_product(sz_abstract_factory: SzAbstractFactory) -> None:
-    """Create SzProduct."""
-    actual = sz_abstract_factory.create_product()
+def test_create_product(sz_abstractfactory: SzAbstractFactory) -> None:
+    """Test SzAbstractFactory.create_product()."""
+    actual = sz_abstractfactory.create_product()
     assert isinstance(actual, SzProduct)
 
 
-def test_reinitialize(sz_abstract_factory: SzAbstractFactory) -> None:
-    """Create SzConfig."""
+def test_help_1(sz_abstractfactory: SzAbstractFactory) -> None:
+    """Test SzAbstractFactory.help()."""
+    sz_abstractfactory.help()
+
+
+def test_help_2(sz_abstractfactory: SzAbstractFactory) -> None:
+    """Test SzAbstractFactory.help(...)."""
+    sz_abstractfactory.help("create_configmanager")
 
     datasources = [f"TEST_DATASOURCE_{datetime.now().timestamp()}"]
 
     # Create Senzing objects.
 
-    sz_config = sz_abstract_factory.create_config()
-    sz_configmanager = sz_abstract_factory.create_configmanager()
+    sz_configmanager = sz_abstractfactory.create_configmanager()
+    sz_config = sz_configmanager.create_config_from_template()
 
-    # Get existing Senzing configuration.
-
-    old_config_id = sz_configmanager.get_default_config_id()
-    old_json_config = sz_configmanager.get_config(old_config_id)
-    config_handle = sz_config.import_config(old_json_config)
-
-    # Add DataSources to existing Senzing configuration.
+    # Add DataSources to Senzing configuration.
 
     for datasource in datasources:
-        sz_config.add_data_source(config_handle, datasource)
+        sz_config.add_data_source(datasource)
 
     # Persist new Senzing configuration.
 
-    new_json_config = sz_config.export_config(config_handle)
-    new_config_id = sz_configmanager.add_config(new_json_config, "Add My datasources")
-    sz_configmanager.replace_default_config_id(old_config_id, new_config_id)
+    config_definition = sz_config.export()
+    config_id = sz_configmanager.set_default_config(config_definition, "Add My datasources")
 
     # Update other Senzing objects.
 
-    sz_abstract_factory.reinitialize(new_config_id)
+    sz_abstractfactory.reinitialize(config_id)
 
 
 # -----------------------------------------------------------------------------
@@ -100,8 +92,8 @@ def test_context() -> None:
     """Test constructor."""
     with SzAbstractFactoryGrpc(grpc_channel=get_grpc_channel()) as actual:
         assert isinstance(actual, SzAbstractFactory)
-        sz_config = actual.create_config()
-        assert isinstance(sz_config, SzConfig)
+        sz_config = actual.create_configmanager()
+        assert isinstance(sz_config, SzConfigManager)
 
 
 # -----------------------------------------------------------------------------
@@ -109,10 +101,10 @@ def test_context() -> None:
 # -----------------------------------------------------------------------------
 
 
-@pytest.fixture(name="sz_abstract_factory", scope="function")
-def sz_abstract_factory_fixture() -> SzAbstractFactory:
+@pytest.fixture(name="sz_abstractfactory", scope="function")
+def szabstractfactory_fixture() -> SzAbstractFactory:
     """
-    Single SzAbstractFactoryGrpc object to use for all tests.
+    SzAbstractFactory object to use for all tests.
     """
     result = SzAbstractFactoryGrpc(**FACTORY_PARAMETERS)
     return result