+++ tests/szconfigmanager_test.py
@@ -1,8 +1,10 @@
+import datetime
 import json
 
 import pytest
 from pytest_schema import Optional, Or, schema
 from senzing import (
+    SzBadInputError,
     SzConfig,
     SzConfigManager,
     SzConfigurationError,
@@ -15,117 +17,139 @@ from senzing_grpc import SzConfigGrpc, SzConfigManagerGrpc
 from .helpers import get_grpc_channel
 
 # -----------------------------------------------------------------------------
-# Testcases
+# Test cases
 # -----------------------------------------------------------------------------
 
 
-def test_add_config(sz_configmanager: SzConfigManager, sz_config: SzConfig) -> None:
-    """Test SzConfigManager().add_config()."""
-    config_handle = sz_config.create_config()
-    config_definition = sz_config.export_config(config_handle)
-    config_comment = "Test"
-    actual = sz_configmanager.add_config(config_definition, config_comment)
-    assert isinstance(actual, int)
-    assert actual > 0
-
-
-def test_add_config_dict(sz_configmanager: SzConfigManager, sz_config: SzConfig) -> None:
-    """Test SzConfigManager().add_config()."""
-    config_handle = sz_config.create_config()
-    config_definition = sz_config.export_config(config_handle)
-    config_definition_as_dict = json.loads(config_definition)
-    config_comment = "Test"
-    actual = sz_configmanager.add_config(config_definition_as_dict, config_comment)
-    assert isinstance(actual, int)
-    assert actual > 0
-
-
-def test_add_config_bad_config_definition_type(
-    sz_configmanager: SzConfigManager,
-) -> None:
-    """Test SzConfigManager().add_config()."""
-    bad_config_definition = 0
-    config_comment = "Test"
-    with pytest.raises(TypeError):
-        sz_configmanager.add_config(bad_config_definition, config_comment)  # type: ignore[arg-type]
-
-
-def test_add_config_bad_config_definition_value(
-    sz_configmanager: SzConfigManager,
-) -> None:
-    """Test SzConfigManager().add_config()."""
-    bad_config_definition = '{"just": "junk"}'
-    config_comment = "Test"
-    actual = sz_configmanager.add_config(bad_config_definition, config_comment)
-    assert isinstance(actual, int)
-    assert actual > 0
-
-
-def test_add_config_bad_config_comment_type(sz_configmanager: SzConfigManager, sz_config: SzConfig) -> None:
-    """Test SzConfigManager().add_config()."""
-    config_handle = sz_config.create_config()
-    config_definition = sz_config.export_config(config_handle)
-    bad_config_comment = 0
-    with pytest.raises(TypeError):
-        sz_configmanager.add_config(config_definition, bad_config_comment)  # type: ignore[arg-type]
-
-
-def test_get_config(sz_configmanager: SzConfigManager) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
+def test_create_config_from_config_id(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.create_config_from_config_id()."""
     config_id = sz_configmanager.get_default_config_id()
-    actual = sz_configmanager.get_config(config_id)
+    sz_config = sz_configmanager.create_config_from_config_id(config_id)
+    actual = sz_config.export()
     actual_as_dict = json.loads(actual)
     assert schema(config_schema) == actual_as_dict
 
 
-def test_get_config_bad_config_id_type(sz_configmanager: SzConfigManager) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
+def test_create_config_from_config_id_bad_config_id_type(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.create_config_from_config_id()."""
     bad_config_id = "string"
     with pytest.raises(TypeError):
-        sz_configmanager.get_config(bad_config_id)  # type: ignore[arg-type]
+        sz_configmanager.create_config_from_config_id(bad_config_id)  # type: ignore[arg-type]
 
 
-def test_get_config_bad_config_id_value(sz_configmanager: SzConfigManager) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
+def test_create_config_from_config_id_bad_config_id_value(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.create_config_from_config_id()."""
     bad_config_id = 1234
     with pytest.raises(SzConfigurationError):
-        sz_configmanager.get_config(bad_config_id)
+        sz_configmanager.create_config_from_config_id(bad_config_id)
+
+
+def test_create_config_from_string(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.create_config_from_config_id()."""
+    config_id = sz_configmanager.get_default_config_id()
+    sz_config = sz_configmanager.create_config_from_config_id(config_id)
+    config_definition = sz_config.export()
+
+    new_sz_config = sz_configmanager.create_config_from_string(config_definition)
+    actual = new_sz_config.export()
+
+    actual_as_dict = json.loads(actual)
+    assert schema(config_schema) == actual_as_dict
+
+
+def test_create_config_from_string_bad_config_definition(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.create_config_from_string()."""
+    bad_config_definition = "}{"
+    with pytest.raises(SzBadInputError):
+        _ = sz_configmanager.create_config_from_string(bad_config_definition)
+
+
+def test_create_config_from_template(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.create_config_from_template()."""
+    sz_config = sz_configmanager.create_config_from_template()
+    actual = sz_config.export()
+    actual_as_dict = json.loads(actual)
+    assert schema(config_schema) == actual_as_dict
 
 
 def test_get_configs(sz_configmanager: SzConfigManager) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
+    """Test SzConfigManager.get_configs()."""
     actual = sz_configmanager.get_configs()
     actual_as_dict = json.loads(actual)
     assert schema(config_list_schema) == actual_as_dict
 
 
 def test_get_default_config_id(sz_configmanager: SzConfigManager) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
+    """Test SzConfigManager.get_default_config_id()."""
     actual = sz_configmanager.get_default_config_id()
     assert isinstance(actual, int)
 
 
-def test_replace_default_config_id(sz_configmanager: SzConfigManager, sz_config: SzConfig) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
-    current_default_config_id = sz_configmanager.get_default_config_id()
-    config_handle = sz_config.create_config()
-    for data_source_code in TRUTHSET_DATASOURCES:
-        sz_config.add_data_source(config_handle, data_source_code)
-    data_source_code = "REPLACE_DEFAULT_CONFIG_ID"
-    sz_config.add_data_source(config_handle, data_source_code)
-    config_definition = sz_config.export_config(config_handle)
+def test_help_1(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.help()."""
+    sz_configmanager.help()
+
+
+def test_help_2(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.help(...)."""
+    sz_configmanager.help("register_config")
+
+
+def test_register_config(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.register_config()."""
+    sz_config = sz_configmanager.create_config_from_template()
+    config_definition = sz_config.export()
     config_comment = "Test"
-    new_default_config_id = sz_configmanager.add_config(config_definition, config_comment)
+    actual = sz_configmanager.register_config(config_definition, config_comment)
+    assert isinstance(actual, int)
+    assert actual > 0
+
+
+def test_register_config_bad_config_definition_type(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.register_config()."""
+    bad_config_definition = 0
+    config_comment = "Test"
+    with pytest.raises(TypeError):
+        sz_configmanager.register_config(bad_config_definition, config_comment)  # type: ignore[arg-type]
+
+
+def test_register_config_bad_config_definition_value(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.register_config()."""
+    bad_config_definition = '{"just": "junk"}'
+    config_comment = "Test"
+    actual = sz_configmanager.register_config(bad_config_definition, config_comment)
+    assert isinstance(actual, int)
+    assert actual > 0
+
+
+def test_register_config_bad_config_comment_type(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.register_config()."""
+    sz_config = sz_configmanager.create_config_from_template()
+    config_definition = sz_config.export()
+    bad_config_comment = 0
+    with pytest.raises(TypeError):
+        sz_configmanager.register_config(config_definition, bad_config_comment)  # type: ignore[arg-type]
+
+
+def test_replace_default_config_id(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.replace_default_config_id()."""
+    current_default_config_id = sz_configmanager.get_default_config_id()
+    sz_config = sz_configmanager.create_config_from_config_id(current_default_config_id)
+    for data_source_code in TRUTHSET_DATASOURCES:
+        sz_config.add_data_source(data_source_code)
+    data_source_code = "REPLACE_DEFAULT_CONFIG_ID"
+    sz_config.add_data_source(data_source_code)
+    config_definition = sz_config.export()
+    config_comment = "Test"
+    new_default_config_id = sz_configmanager.register_config(config_definition, config_comment)
     assert current_default_config_id != new_default_config_id
     sz_configmanager.replace_default_config_id(current_default_config_id, new_default_config_id)
     actual = sz_configmanager.get_default_config_id()
     assert actual == new_default_config_id
 
 
-def test_replace_default_config_id_bad_new_default_config_id_type(
-    sz_configmanager: SzConfigManager,
-) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
+def test_replace_default_config_id_bad_new_default_config_id_type(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.replace_default_config_id()."""
     current_default_config_id = sz_configmanager.get_default_config_id()
     bad_new_default_config_id = "string"
     with pytest.raises(TypeError):
@@ -134,67 +158,79 @@ def test_replace_default_config_id_bad_new_default_config_id_type(
         )
 
 
-def test_replace_default_config_id_bad_new_default_config_id_value(
-    sz_configmanager: SzConfigManager,
-) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
+def test_replace_default_config_id_bad_new_default_config_id_value(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.replace_default_config_id()."""
     current_default_config_id = sz_configmanager.get_default_config_id()
     bad_new_default_config_id = 1234
     with pytest.raises(SzConfigurationError):
         sz_configmanager.replace_default_config_id(current_default_config_id, bad_new_default_config_id)
 
 
-def test_replace_default_config_id_bad_current_default_config_id_type(
-    sz_configmanager: SzConfigManager, sz_config: SzConfig
-) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
+def test_replace_default_config_id_bad_current_default_config_id_type(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.replace_default_config_id()."""
     bad_current_default_config_id = "string"
-    config_handle = sz_config.create_config()
+    sz_config = sz_configmanager.create_config_from_template()
     data_source_code = "REPLACE_DEFAULT_CONFIG_ID"
-    sz_config.add_data_source(config_handle, data_source_code)
-    config_definition = sz_config.export_config(config_handle)
+    sz_config.add_data_source(data_source_code)
+    config_definition = sz_config.export()
     config_comment = "Test"
-    new_default_config_id = sz_configmanager.add_config(config_definition, config_comment)
+    new_default_config_id = sz_configmanager.register_config(config_definition, config_comment)
     with pytest.raises(TypeError):
         sz_configmanager.replace_default_config_id(
             bad_current_default_config_id, new_default_config_id  # type: ignore[arg-type]
         )
 
 
-def test_replace_default_config_id_bad_current_default_config_id_value(
-    sz_configmanager: SzConfigManager, sz_config: SzConfig
-) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
+def test_replace_default_config_id_bad_current_default_config_id_value(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.replace_default_config_id()."""
     bad_current_default_config_id = 1234
-    config_handle = sz_config.create_config()
+    sz_config = sz_configmanager.create_config_from_template()
     data_source_code = "CUSTOMERS"
-    sz_config.add_data_source(config_handle, data_source_code)
-    config_definition = sz_config.export_config(config_handle)
+    sz_config.add_data_source(data_source_code)
+    config_definition = sz_config.export()
     config_comment = "Test"
-    new_default_config_id = sz_configmanager.add_config(config_definition, config_comment)
+    new_default_config_id = sz_configmanager.register_config(config_definition, config_comment)
     with pytest.raises(SzReplaceConflictError):
         sz_configmanager.replace_default_config_id(bad_current_default_config_id, new_default_config_id)
 
 
-def test_set_default_config_id(sz_configmanager: SzConfigManager, sz_config: SzConfig) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
+def test_set_default_config(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.set_default_config()."""
     old_config_id = sz_configmanager.get_default_config_id()
-    config_handle = sz_config.create_config()
-    data_source_code = "CUSTOMERS"
-    sz_config.add_data_source(config_handle, data_source_code)
-    config_definition = sz_config.export_config(config_handle)
+    sz_config = sz_configmanager.create_config_from_config_id(old_config_id)
+    data_source_code = "TEST_DATASOURCE_" + datetime.datetime.now(datetime.timezone.utc).isoformat()
+    sz_config.add_data_source(data_source_code)
+    config_definition = sz_config.export()
     config_comment = "Test"
-    config_id = sz_configmanager.add_config(config_definition, config_comment)
+    actual = sz_configmanager.set_default_config(config_definition, config_comment)
+    assert actual > 0
+
+
+def test_set_default_config_bad_config_definition(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.set_default_config()."""
+    bad_config_definition = "}{"
+    config_comment = "Test"
+    with pytest.raises(SzConfigurationError):
+        _ = sz_configmanager.set_default_config(bad_config_definition, config_comment)
+
+
+def test_set_default_config_id(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.set_default_config_id()."""
+    old_config_id = sz_configmanager.get_default_config_id()
+    sz_config = sz_configmanager.create_config_from_template()
+    data_source_code = "CUSTOMERS"
+    sz_config.add_data_source(data_source_code)
+    config_definition = sz_config.export()
+    config_comment = "Test"
+    config_id = sz_configmanager.register_config(config_definition, config_comment)
     assert old_config_id != config_id
     sz_configmanager.set_default_config_id(config_id)
     actual = sz_configmanager.get_default_config_id()
     assert actual == config_id
 
 
-def test_set_default_config_id_bad_config_id_type(
-    sz_configmanager: SzConfigManager,
-) -> None:
-    """Test SzConfigManager().get_default_config_id()."""
+def test_set_default_config_id_bad_config_id_type(sz_configmanager: SzConfigManager) -> None:
+    """Test SzConfigManager.set_default_config_id()."""
     bad_config_id = "string"
     with pytest.raises(TypeError):
         sz_configmanager.set_default_config_id(bad_config_id)  # type: ignore[arg-type]
@@ -203,7 +239,7 @@ def test_set_default_config_id_bad_config_id_type(
 def test_set_default_config_id_bad_config_id_value(
     sz_configmanager: SzConfigManager,
 ) -> None:
-    """Test SzConfigManager().set_default_config_id()."""
+    """Test SzConfigManager.set_default_config_id()."""
     bad_config_id = 1
     with pytest.raises(SzConfigurationError):
         sz_configmanager.set_default_config_id(bad_config_id)
@@ -224,7 +260,8 @@ def test_context_managment() -> None:
     """Test the use of SzConfigManagerTest in context."""
     with SzConfigManagerGrpc(grpc_channel=get_grpc_channel()) as sz_configmanager:
         config_id = sz_configmanager.get_default_config_id()
-        actual = sz_configmanager.get_config(config_id)
+        sz_config = sz_configmanager.create_config_from_config_id(config_id)
+        actual = sz_config.export()
         actual_as_dict = json.loads(actual)
         assert schema(config_schema) == actual_as_dict
 
@@ -237,7 +274,7 @@ def test_context_managment() -> None:
 @pytest.fixture(name="sz_config", scope="function")
 def szconfig_fixture() -> SzConfig:
     """
-    Single szconfigmanager object to use for all tests.
+    SzConfig object to use for all tests.
     """
     result = SzConfigGrpc(grpc_channel=get_grpc_channel())
     return result
@@ -245,8 +282,7 @@ def szconfig_fixture() -> SzConfig:
 
 @pytest.fixture(name="sz_configmanager", scope="function")
 def szconfigmanager_fixture() -> SzConfigManager:
-    """Single szconfigmanager object to use for all tests.
-    build_engine_vars is returned from conftest.pys"""
+    """SzCconfigManager object to use for all tests."""
     result = SzConfigManagerGrpc(grpc_channel=get_grpc_channel())
     return result